fix(quality): address SonarCloud + DeepSource findings from first CI pass
Three findings from the first CI run on this branch:

1. SonarCloud python:S7497 (BUG, MAJOR) on backend/app/api.py:102 — the
   try/except CancelledError block did not re-raise. In this specific
   context we *are* the cancellation originator (the preceding line is
   cleanup_task.cancel()), so propagating the CancelledError would be
   wrong. Switched to `await asyncio.gather(cleanup_task,
   return_exceptions=True)`, which awaits the task and treats any
   resulting exception (including the CancelledError we just triggered)
   as a returned value. CodeQL reads this as clearly effectful and
   Sonar no longer sees a CancelledError being silently swallowed.

2. SonarCloud typescript:S3735 + DeepSource JS-0098 on the three
   controller useEffect hooks — Sonar flags `void loadEvent()` etc. as
   unnecessary. Since loadEvent / loadProfile / loadEvent (form) all
   handle their own errors internally with try/except, the promise can
   never reject, so there is no unhandled-rejection risk. Dropped the
   `void` prefix; the bare call is fine under our eslint config (no
   no-floating-promises rule) and both linters are happy.

3. SonarCloud "11.9% duplication on new code" on the test suite —
   ui/tests/coverage-closure-analytics.test.tsx duplicated the hoisted
   mock block + vi.mock() calls from coverage-closure-regressions.test
   .tsx. Merged the three new tests (EventsPage media-change handler,
   EventsPage analytics-rejection, EventDetail analytics-rejection)
   into coverage-closure-regressions.test.tsx and deleted the separate
   file. Added `getEvents`/`getFavorites` to the shared eventServiceMock
   and three helper functions (mountMatchMediaMock, renderEventsPage,
   renderEventDetailProbe) to keep the new tests concise.
